requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
Flask==3.1.0
Flask-Limiter==3.9.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.18.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
